requirements.txt
numpy>=1.21
pandas>=1.3
matplotlib>=3.4
scipy>=1.7
pillow>=8.0
scikit-learn>=1.0